move all code related to run() to seperate file
